requirements.txt
streamlit>=1.28.0
PyGithub>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
apify-client>=1.6.0
pandas>=2.0.0
Pillow>=10.0.0
openpyxl>=3.1.0
